Fix compilation in 64 bit architecture

+++ src/graph/misc/graph_rewiring.cc
@@ -32,7 +32,7 @@ using namespace boost::lambda;
 void random_rewire(GraphInterface& gi, string strat, bool self_loops,
                    bool parallel_edges, size_t seed)
 {
-    rng_t rng(seed);
+    rng_t rng((uint32_t) seed);
 
     GraphInterface::edge_index_map_t edge_index =
         any_cast<GraphInterface::edge_index_map_t>(gi.GetEdgeIndex());
